updater: Return extra code when AppCommands fail to launch/time out, log
A ping with the value `kEventAppCommandComplete` = `41` is sent if
usagestats are enabled after an app command completes execution.

If the app command launched successfully, the result returned by the app
command process will be reported in `error` in the ping.

If the app command fails to launch, code `kErrorAppCommandLaunchFailed `
is reported in the `extra_code1` in the ping, along with the actual
error code that caused that launch failure in `error`.

If the app command times out, code `kErrorAppCommandTimedOut` is
reported in the `extra_code1` in the ping, along with the error code
`HRESULT_FROM_WIN32(ERROR_TIMEOUT)` in `error`.

Bug: 324271167
Change-Id: I1ec8ccb9803fb4a8116e071051f8658a6e5b44cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5354158
Auto-Submit: S Ganesh <ganesh@chromium.org>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1270384}