[IP Protection] Add enterprise policy to disable feature
To provide enterprise admins with a way to opt-out of IP Protection for
their managed clients, this CL implements an enterprise policy that
allows them to disable IP Protection and prevent users from enabling it
via settings (and also enable the feature for their users).

The enterprise policy ties in to the existing kIpProtectionEnabled
pref which will be UI controlled as well. Setting this pref via
enterprise policy should cause the UI code to automatically detect that
the setting is managed and update the UI accordingly (showing a "managed"
indicator).

A high-level design doc (Googlers-only) for this change is at:
https://docs.google.com/document/d/1XmFZzZ2LBXYVT2TpoiDSSaWWsZysB3wazp8Tch7E0HM/edit?usp=sharing&resourcekey=0-O7XmkXCh9dtgdqb0r69rnw

Fixed: 41496985
Change-Id: I82204f09767df4f0c4ad02b0aec7aef9864b7a00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5318549
Commit-Queue: Andrew Williams <awillia@chromium.org>
Reviewed-by: Sébastien Lalancette <seblalancette@chromium.org>
Reviewed-by: Fiona Macintosh <fmacintosh@google.com>
Reviewed-by: Artem Sumaneev <asumaneev@google.com>
Cr-Commit-Position: refs/heads/main@{#1270403}