[UseRecordedBoundsForTiling] Don't append tiles out of recorded bounds
The situation happens when the tiling rect is snapped to be bigger
than the recorded bounds, and CoverageIterator returns "missing" tiles
to cover the empty area. The tiles should be ignored, otherwise they
would be mistakenly treated as checkerboarded and drawn with the
safe background color.

Bug: 41490692
Change-Id: Iab2f3a29cc55b4264385470368d6d3e958f6bf6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5353406
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1270440}