[ViewsAX] Remove WebAXPlatformTreeManagerDelegate
This brings us yet another step towards moving BAM-related
classes to ui/ to allow them to be able to be used for views
as well.
This change is mostly a straightforward rename from
WebAXPlatformTreeManagerDelegate to ui::AXPlatformTreeManagerDelegate,
but also includes the following notable changes:

* Removal of debug check in BAM::GetParentManager
* Added methods to AXPlatformTreeManagerDelegate:
  * AccessibilityIsRootFrame
  * IsChromeNewTabPage
If in a Views context, returning false/nullptr will still
give us the intended result, so I don't think this
should hamper the goal of moving BAM-related code to ui/

Additionally, the following method was also added for android
compatibility, where there is no concept of views:
* AccessibilityGetWebContentsAccessibility
This is method is tied to the RenderFrameHostImpl's ability
to get the RenderWidgetHostViewBase, making this difficult
to decouple from the RenderWidgetHostViewBase
At least for now, this is moved to AXPlatformTreeManagerDelegate.

Bug: 40672441
Change-Id: I6284cb56cc52b3ac4996e5a47095beb7ad24881b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5326217
Auto-Submit: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Jacques Newman <janewman@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1270465}