WebUI: Move cr-searchable-drop-down to only user
Move cr-searchable-drop-down to Print Preview. This element is only
used in CrOS Print Preview, and is not allowed on other platforms due
to use of deprecated iron-dropdown. Moving it out of the shared
resources may help to discourage any further use.

Rename it to searchable_drop_down_cros so that the CrOS Printing team
can own it.

Bug: 40943652
Change-Id: Ifeb79a69e85fefc1606abbd9116e126cd6f43dfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5356435
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1270489}