d3d11: specify texture format and view format in DawnTextureInfo
Skia needs the texture format and aspect to determine the depth buffer
size. So we have to pass the texture format in DawnTextureInfo.
This CL only set fViewFormat in DawnTextureInfo for creating promise
image for a multiplanar shared image, and when the promise image is
fulfilled, fFormat and fAspect in DawnTextureInfo will be set.
In future, skia graphite may determine view format based on fFormat
and fAspect.

Bug: 324422644
Change-Id: I7bee4cd918e5a6a3fe353ac2373919ded442d6f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5348918
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Peng Huang <penghuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1270547}