Change PersistentRepeatingTimer to RepeatingTimer in PlusAddressService.
Currently, PlusAddressService clears the pref for when the plus
address server was polled the last time on creation. As a result, it
(almost) works identically to a RepeatingTimer. The only difference
is on user signout and signin again: In this case, the
PersistentRepeatingTimer did not poll the manager again immediately.
However, arguably, this is the right thing to do: Otherwise, a freshly
signed in user may not receive their plus addresses for a while.

The CL removes the pref needed for the PersistentRepeatingTimer.
However, it does not remove the PrefService pointer from the
PlusAddressService, nor does it change the tests. We expect there to be
new prefs soon, e.g. for enterprise policies.

Bug: b/322279583
Change-Id: I655ea9f81d9ebf91da44a906e6b6a4ff3314126a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5355751
Reviewed-by: Dominic Battre <battre@chromium.org>
Reviewed-by: Matt Reichhoff <mreichhoff@chromium.org>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1270554}